% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_histogram.R
\name{optimal_histogram}
\alias{optimal_histogram}
\title{Optimal Histogram}
\usage{
optimal_histogram(x, axis, ...)
}
\arguments{
\item{x}{A list containing values belonging to the following data types: numeric, integer or NA. This parameter is named x because it represents the numerical variable that will be plotted on the x-axis of this histogram.}

\item{axis}{An object of class "character", indicating the x-axis label. This parameter is named axis because its input is used create the label of the x-axis of the histogram.}

\item{...}{Other parameters to pass to geom_histogram()}
}
\value{
A histogram plot that visualizes the distribution of x
}
\description{
This function creates a histogram with the optimal bin-width, as calculated by the Freedman-Diaconis rule. It is a wrapper for geom_histogram().
}
\examples{
optimal_histogram(c(2, 3, 3, 3, 4, 4, 2, 2, 5, 5, 5, 5, 3), "Numbers")
optimal_histogram(c(2, 3, 3, 3, 5, 4, 2, 3, 2, 5, 4, 5, 1), "Mean Neutrophil Counts", fill = "darkgoldenrod4", color = "darkgoldenrod3", alpha = 0.5)
}
